Add IpStatsRepo for IP management and statistics handling
- Implemented methods for checking and managing IP bans, including ban overrides and detailed ban information retrieval.
- Added functionality for tracking IPs, updating IP statistics, and managing category changes with history logging.
- Introduced pagination for retrieving attackers and tracked IPs, along with methods for exporting IPs based on categories.
- Enhanced error handling and logging throughout the repository methods.
